requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
python-multipart>=0.0.9
aiofiles>=24.1.0
pydantic-settings>=2.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
